Enable the TextEditor on the response view so user can scroll and refactoring.
